+++ Nos/Views/NoteComposer/ImagePickerButton.swift
@@ -3,13 +3,19 @@ import SwiftUI
 import Dependencies
 
 struct ImagePickerButton<Label>: View where Label: View {
-
+    
+    @Dependency(\.analytics) private var analytics
+    @Dependency(\.featureFlags) private var featureFlags
+    
     /// The device to be used initially when the user chooses to use the camera.
     let cameraDevice: UIImagePickerController.CameraDevice
     /// The types of content the user can choose.
     let mediaTypes: [UTType]
     let onCompletion: ((URL) -> Void)
     let label: () -> Label
+    
+    @State
+    private var selectedContentURL: URL?
 
     /// State used to present or hide a confirmation dialog that lets the user select the ImagePicker source.
     @State
@@ -18,12 +24,13 @@ struct ImagePickerButton<Label>: View where Label: View {
     /// State used to present or hide an alert that lets the user go to settings.
     @State
     private var showSettingsAlert = false
+    
+    @State
+    private var showSensitiveContentAlert = false
 
     /// Source used by ImagePicker when opening it
     @State
     private var imagePickerSource: UIImagePickerController.SourceType?
-    
-    @Dependency(\.analytics) private var analytics
     
     @EnvironmentObject private var router: Router
 
@@ -108,6 +115,23 @@ struct ImagePickerButton<Label>: View where Label: View {
                 Text("openSettingsMessage", tableName: "ImagePicker")
             }
         )
+        .alert(
+            "contentWarning",
+            isPresented: $showSensitiveContentAlert,
+            actions: {
+                Button("cancel") {}
+                Button("upload") {
+                    if let selectedContentURL {
+                        onCompletion(selectedContentURL)
+                    } else {
+                        assertionFailure("ImagePickerButton did not have the content URL")
+                    }
+                }
+            },
+            message: {
+                Text("sensitiveContentUploadWarning")
+            }
+        )
         .sheet(isPresented: showImagePicker) {
             ImagePickerUIViewController(
                 sourceType: imagePickerSource ?? .photoLibrary, 
@@ -129,13 +153,27 @@ struct ImagePickerButton<Label>: View where Label: View {
     /// Called when a user chooses an image or video.
     /// - Parameter url: The URL of the image or video on disk.
     func userPicked(url: URL?) {
-        if let url {
-            analytics.selectedImage()
-            onCompletion(url)
-        } else {
-            analytics.cancelledImageSelection()
+        imagePickerSource = nil // dismisses the picker
+        
+        Task {  // The separate Task allows the picker view to dismiss before we potentially show another pop-up.
+            if let url {
+                analytics.selectedImage()
+                
+                if featureFlags.isEnabled(.sensitiveContentAnalysisEnabled) {
+                    let shouldWarn = await SensitiveContentController.shared.shouldWarnUserUploadingFile(at: url)
+                    if shouldWarn {
+                        selectedContentURL = url
+                        showSensitiveContentAlert = true
+                    } else {
+                        onCompletion(url)
+                    }
+                } else {
+                    onCompletion(url)
+                }
+            } else {
+                analytics.cancelledImageSelection()
+            }
         }
-        imagePickerSource = nil
     }
 }